requirements.txt
# Core ML and CV dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Dataset and model utilities
datasets>=2.14.0
transformers>=4.30.0
huggingface-hub>=0.16.0

# Web and API
fastapi>=0.100.0
websockets>=11.0.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Database - Supabase removed for Kaggle dataset integration
# supabase>=1.0.0  # Removed for Kaggle dataset integration
python-dotenv>=1.0.0  # Keeping for local environment variables

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0